[evas_events.c] nicer do !! once.
[evas_map.c] 1 bug-- ... when maps got enabled and disabled you'd find in softwarre
that you'd get dirty areas that didnt re-rneder. fixed with nasty hack
- force a full update if anyone in the evas object tree changed maps
from on to off and vice-versa

[evas_object_main.c] !! once

[evas_render.c] fix nasty uncaught clip + map bug! fix fix fix!!!!!

[evas_font_draw.c] Do kerning correctly for word cache.
Technically this should speed it up a little. That's a nice change for a bug
fix ;-)

[Evas_Engine_DirectFB.h] Fix header include
The current way of including directfb.h makes it hard for people with multiple
versions of DirectFB installed.
This is the correct way to include that header, and it's the way
ecore_directfb does it.
Patch by: Eduardo Felipe <eduardofelipe87@>