Evas textblock: Fixed range_text_get with TEXT_PLAIN.
Until now, it only supported MARKUP, now it also supports plain.
Haven't tested it much, but it seems to be working. It can help a lot to
people who want to analyse the text, for example when implementing a spell
checker, or "search", as now the text maps exactly like the cursors map.
Still have to decide if that's the wanted API and set it in stone, or not
promise this and expose a different API for these kind of tasks.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/evas@62011 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33