evas evas_object_smart.c: Reverted smart callback calling sequence. Now, smart callback is prepended among the same priorities. After the discussion with hermet and raster.
I fixed this because I think it's better keeping the default behavior
even we have new priority option. This is for evas backward
compatibility and for the world peace. Although I know there is no
official policy and it's not recommended to rely on the smart callback
calling sequence in application, I really hope evas needs to have
compatibility. Otherwise we need to adopt the new policy in next
version of evas.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/evas@62792 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33