Cleaned up some evas_gl code and added surface cap feature.
Previously, evas_gl_surface_create() didn't actually do 
the render buffer attach to the the FBO. It was performed when 
the make_current was called for the first time.  The issue 
was that even though the surface was successfully created with 
the given configuration, there was a possibility of make_current 
failing with the error message "FBO not complete" because of 
the surface configuration.

So, I've added a piece of code that checks the FBO 
capabilities beforehand to set up a available surface configurations
so that it doesn't have to fail during make_current for unsupported 
surface format.

Also, I've changed the surface config in a way that once the 
user calls evas_gl_surface_create(), evas gl sets the config
parameter with configuration that evas_gl is actually using.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/evas@70617 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33